Añadiendo el LawnMover

+++ src/domain/LawnMover.java
@@ -16,13 +16,23 @@ public class LawnMover implements Thing{
 
     @Override
     public void update() throws PoobVsZombiesException {
+        if(!isAlive()){
+            poobVsZombies.addLawnMoverToRemove(this);
+            return;
+        }
+        long currentTime = System.currentTimeMillis();
         if(modeAtack()){
-            long currentTime = System.currentTimeMillis();
+            System.out.println(currentTime - lastDeathTime);
             if ((currentTime - lastDeathTime) >= COOLDOWN) {
                 deleteZombies();
                 move();
                 lastDeathTime = currentTime;
             }
+        }
+        if (isActive()&&(currentTime - lastDeathTime) >= COOLDOWN) {
+            System.out.println(currentTime - lastDeathTime);
+            move();
+            lastDeathTime = currentTime;
         }
     }
 
@@ -38,11 +48,11 @@ public class LawnMover implements Thing{
 
     private void move(){
         if(y<poobVsZombies.getBoard()[0].length){
-            y += 1;
+            poobVsZombies.addLawnMoverToMove(this);
         }
     }
 
-    private boolean modeAtack(){
+    public boolean modeAtack(){
         for(Thing thing :poobVsZombies.getBoard()[x][y]){
             if(thing instanceof Zombie){
                 return true;
@@ -51,10 +61,14 @@ public class LawnMover implements Thing{
         return false;
     }
 
+    public boolean isActive(){
+        return y != 0;
+    }
+
 
     @Override
     public boolean isAlive() {
-        if(y>poobVsZombies.getBoard()[0].length){
+        if(y==poobVsZombies.getBoard()[0].length-1){
             return false;
         }
         return true;
@@ -62,6 +76,22 @@ public class LawnMover implements Thing{
 
     @Override
     public void takeDamage(int damage) {
-
     }
+
+    public int getX() {
+        return x;
+    }
+
+    public int getY() {
+        return y;
+    }
+
+    public void setX(int x){
+        this.x = x;
+    }
+
+    public void setY(int y){
+        this.y = y;
+    }
+
 }
